Upgrade to Next 16 / Node 24, move API routes off the deprecated edge runtime
Vercel deprecated Edge Functions in favor of Fluid compute, and Next 13
left support long ago. Pages Router edge API routes move to App Router
route handlers on the default Node runtime; the download route's
NextFetchEvent.waitUntil becomes after() from next/server. Handler logic
and response shapes are unchanged, and the engines field pins Node 24.x
for Vercel builds.

random-words v1 silently ignored the seed and wordsPerString options this
code passes; v2 honors them, so lexica's fake pagination is now actually
deterministic per page.

The repo carried both yarn.lock and package-lock.json; yarn.lock is gone
so installs resolve one way. next lint was removed in Next 16, so linting
is now eslint 9 flat config via eslint-config-next's exported presets
(which flagged the ported photo: any — now Record<string, unknown>).
