+++ app/layout.tsx
@@ -1,0 +1,15 @@
+export const metadata = {
+    title: 'unsplash-api',
+}
+
+export default function RootLayout({
+    children,
+}: {
+    children: React.ReactNode
+}) {
+    return (
+        <html lang="en">
+            <body>{children}</body>
+        </html>
+    )
+}
