+++ eslint.config.mjs
@@ -1,0 +1,4 @@
+import coreWebVitals from 'eslint-config-next/core-web-vitals'
+import typescript from 'eslint-config-next/typescript'
+
+export default [coreWebVitals, typescript].flat()
