+++ package.json
@@ -2,24 +2,27 @@
     "name": "unsplash-api",
     "version": "0.1.0",
     "private": true,
+    "engines": {
+        "node": "24.x"
+    },
     "scripts": {
         "dev": "next dev",
         "build": "next build",
         "start": "next start",
-        "lint": "next lint"
+        "lint": "eslint ."
     },
     "dependencies": {
-        "next": "^13.3.1",
-        "random-words": "^1.3.0",
-        "react": "18.2.0",
-        "react-dom": "18.2.0"
+        "next": "^16.3.1",
+        "random-words": "^2.0.1",
+        "react": "^19.2.0",
+        "react-dom": "^19.2.0"
     },
     "devDependencies": {
-        "@types/node": "18.15.13",
-        "@types/react": "18.0.38",
-        "@types/react-dom": "18.0.11",
-        "eslint": "8.39.0",
-        "eslint-config-next": "^13.3.1",
-        "typescript": "5.0.4"
+        "@types/node": "^24",
+        "@types/react": "^19",
+        "@types/react-dom": "^19",
+        "eslint": "^9",
+        "eslint-config-next": "^16.3.1",
+        "typescript": "^5.9.3"
     }
 }
